+++ projects/client-ws/package.json
@@ -1,11 +1,13 @@
 {
   "name": "@remixproject/plugin-ws",
-  "version": "0.0.1",
+  "version": "0.0.2-alpha.4",
   "description": "Library to build Plugin over Websocket. Used by Remix IDE",
-  "contributors": [{
-    "name": "GrandSchtroumpf",
-    "email": "grandschtroumpf@ethereum.org"
-  }],
+  "contributors": [
+    {
+      "name": "GrandSchtroumpf",
+      "email": "grandschtroumpf@ethereum.org"
+    }
+  ],
   "dependencies": {
     "@remixproject/plugin": "^0.1.18"
   },
