+++ projects/client/src/iframe.ts
@@ -1,6 +1,6 @@
 import { PluginDevMode, PluginClient, PluginOptions } from './client'
-import { Message, PluginApi, ApiMap, ProfileMap, Api, listenEvent, callEvent, RemixApi } from '../../utils'
-import { getApiMap, listenOnThemeChanged } from './api'
+import { Message, PluginApi, ApiMap, ProfileMap, Api, listenEvent, callEvent, RemixApi, Theme } from '../../utils'
+import { getApiMap } from './api'
 
 /** Fetch the default origins for remix */
 export async function getDefaultOrigins() {
@@ -21,6 +21,26 @@ export async function getAllOrigins(devMode: Partial<PluginDevMode> = {}): Promi
     : []
   const defaultOrigins = await getDefaultOrigins()
   return [ ...defaultOrigins, ...localhost, ...devOrigins]
+}
+
+
+/** Start listening on theme changed */
+export async function listenOnThemeChanged(client: PluginClient<any, any>, options?: Partial<PluginOptions<any>>) {
+  if (options && options.customTheme) return
+  const cssLink = document.createElement('link')
+  cssLink.setAttribute('rel', 'stylesheet')
+  document.head.prepend(cssLink)
+  client.onload(async () => {
+    client.on('theme', 'themeChanged', (_theme: Theme) => setTheme(cssLink, _theme))
+    const theme = await client.call('theme', 'currentTheme')
+    setTheme(cssLink, theme)
+  })
+  return cssLink
+}
+
+function setTheme(cssLink: HTMLLinkElement, theme: Theme) {
+  cssLink.setAttribute('href', theme.url)
+  document.documentElement.style.setProperty('--theme', theme.quality)
 }
 
 /**
